+++ bastion/provider.go
@@ -52,8 +52,9 @@ func Provider() *schema.Provider {
 			},
 		},
 		DataSourcesMap: map[string]*schema.Resource{
-			"wallix-bastion_domain":  dataSourceDomain(),
-			"wallix-bastion_version": dataSourceVersion(),
+			"wallix-bastion_configoption": dataSourceConfigoption(),
+			"wallix-bastion_domain":       dataSourceDomain(),
+			"wallix-bastion_version":      dataSourceVersion(),
 		},
 		ResourcesMap: map[string]*schema.Resource{
 			"wallix-bastion_application":                           resourceApplication(),